Check if config file is present.
As explained in the tutorial, config.py.in should be renamed to config.py. If a
user failes to do this, the following error is raised:

   from config import *
   ModuleNotFoundError: No module named 'config'

The deceptive nature of this errors description makes it quite difficult to troubleshoot
at first sight - the module is there, it's the file we're lacking.