Use a blown-up, blurred & darkened photo for background
When the photo does not fit the screen, it gets recaled. Unless the photo has the same imgRatio as the screen, this
results in black bars either at the left/right (when portrait) or above/below (landscape) of the photo. Although it
does the trick, it looks kind of ugly.

Here, the background is generated from the shown image. It gets blown up, blurred and darkened to put the emphasis
still on the "true" picture, wich is the rescaled one. The result is nothing else but slick.

Specific values for the blow-up, blurring and darkening, might need some tweaking.